# Add pigz for parallel decompression by containerd