--- a/packages/containerd/containerd.spec
+++ b/packages/containerd/containerd.spec
@@ -38,6 +38,7 @@ Source1000: clarify.toml
 BuildRequires: git
 BuildRequires: %{_cross_os}glibc-devel
 Requires: %{_cross_os}runc
+Requires: %{_cross_os}pigz
 
 %description
 %{summary}.