--- a/packages/pigz/pkg.rs
+++ b/packages/pigz/pkg.rs
@@ -1,0 +1,1 @@
+// not used