<!--
Tips:
- Please read CONTRIBUTING.md to understand our process and our requests for PRs.
- Please file an issue before creating a PR so we can discuss the change and confirm it's not already being worked on.
-->

**Issue number:**

Closes #2886

**Description of changes:**

Let containerd use pigz (or more precisely, `unpigz`).
pigz doesn't have a LICENSE file (https://github.com/madler/pigz/issues/89) and I decided to use `README`...

If there is no pigz, and the log level of containerd is set to debug, the following log will be output at the time of pull.
```
Mar 14 09:30:08 10.0.2.15 containerd[1175]: time="2023-03-14T09:30:08.609765577Z" level=debug msg="unpigz not found, falling back to go gzip" error="exec: \"unpigz\": executable file not found in $PATH"
```
I have confirmed that with this change that log disappears (i.e. containerd uses unpigz).

Here is a quick benchmark.

machine spec:
Google Compute Engine c3-highcpu-22 Ubuntu 22.04.2 (enable nested virtualization)

Command to execute:
```bash
ctr image pull docker.io/library/mysql:8.0.32
```

without pigz(commit b1ac8ef6922e776ee410b6e3020ec8133c51a959)
```
6.139457575s
7.236172524s
7.730825494s
```

with pigz(this commit)
```
5.842488681s
6.319545383s
5.777997267s
```

**Testing done:** yes

**Terms of contribution:**

By submitting this pull request, I agree that this contribution is dual-licensed under the terms of both the Apache License, version 2.0, and the MIT license.